+++ src/GUI/VirtualDesktopScreen.cs
@@ -25,60 +25,62 @@ namespace Switchie
             Graphics g = e.Graphics;
 
             Window[] windows;
-            windows = Form.Windows.Where(x => x.VirtualDesktopIndex == VirtualDesktop.VirtualDesktopIndex).OrderBy(x => x.ZOrder).ToArray();
 
-            /*
+            // For windows render mode, all windows have to be sorted in z-order to draw from back to front
             if (Form.WindowRenderMode == MainForm.RenderMode.Windows)
             {
-                
+                windows = Form.Windows.Where(x => x.VirtualDesktopIndex == VirtualDesktop.VirtualDesktopIndex).OrderBy(x => x.ZOrder).ToArray();
             }
+
+            // Otherwise order should stay the same (using process id)
             else
             {
-                windows = Form.Windows.Where(x => x.VirtualDesktopIndex == VirtualDesktop.VirtualDesktopIndex).OrderByDescending(x => x.ZOrder).ToArray();
-            }*/
+                var windowsEnumerated = Form.Windows.Where(x => x.VirtualDesktopIndex == VirtualDesktop.VirtualDesktopIndex);
 
+                // Order by process is similar to order is taskbar, as long as the user doesn't move them
+                windows = windowsEnumerated.OrderBy(x => x.ProcessID).ToArray();
+            }
+
+            if (windows.Length == 0) return;
+
+            // Parition the screen equally for the Icons
             int hSpace = Size.Width / windows.Length;
-            int vSpace = Size.Height / windows.Length;
-
-            // Renders the thumbnails of all windows for all desktops from back to front
-            // -> TODO: Different rendering modes, like when all windows are maximized (on smaller screens typical)
-            /*
-            WINDOWPLACEMENT placement = new WINDOWPLACEMENT();
-            placement.length = Marshal.SizeOf(placement);
-            GetWindowPlacement(w.Handle, ref placement);
-            */
-
-            int wnum = windows.Length - 1;
-            foreach (var w in windows)
+            
+            if(hSpace < windows[0].Icon.Width)
             {
-                Color fillColor = w.IsActive ? Form.ActiveWindowColor : Form.WindowColor;
-                Color borderColor = w.IsActive ? Form.ActiveWindowBorderColor : Form.WindowBorderColor;
+                Debug.WriteLine("Not enough Space!");
+            }
+            
+            //int vSpace = Size.Height / windows
+            int wnum = windows.Length - 1;
+            foreach (var wnd in windows)
+            {
+                Color fillColor = wnd.IsActive ? Form.ActiveWindowColor : Form.WindowColor;
+                Color borderColor = wnd.IsActive ? Form.ActiveWindowBorderColor : Form.WindowBorderColor;
 
-                if (Screen.FromHandle(w.Handle).DeviceName == AttachedScreen.DeviceName)
+                if (Screen.FromHandle(wnd.Handle).DeviceName == AttachedScreen.DeviceName)
                 {
                     if (Form.WindowRenderMode == MainForm.RenderMode.Icons)
                     {
-                        // TODO: Window selection doesn't work in this mode 
-                        // TODO: There should to make it looking like in the original (user preferences)
-                        // Window rectangle fill
-                        
-                        // Rectangle for the selected icon, currently fills to whole height
-                        /*g.FillRectangle(
-                            new SolidBrush(fillColor),
-                            new Rectangle(Location.X + wnum * hSpace, 0, hSpace, Size.Height));*/
+                        var yPos = (Size.Height / 2) - wnd.Icon.Height / 2;
+                        var area = new Rectangle(Location.X + wnum * hSpace, yPos-2, hSpace, wnd.Icon.Height+4);
+                        WindowAreas[wnd.Handle] = area;
 
-                        g.DrawImage(w.Icon, new Point(
-                            Location.X + wnum * hSpace + hSpace / 2 - w.Icon.Width / 2,
-                            (Size.Height / 2) - w.Icon.Height / 2
-                        ));
+                        if (wnd.IsActive)
+                        {
+                            // Rectangle for the selected icon, currently fills to whole height
+                            g.FillRectangle(new SolidBrush(fillColor), area);
+                        }
+
+                        g.DrawImage(wnd.Icon, new Point(area.X + hSpace / 2 - wnd.Icon.Width / 2, yPos));
                     }
                     else
                     {
-                        var x = w.Dimensions.X;
-                        var y = w.Dimensions.Y;
+                        var x = wnd.Dimensions.X;
+                        var y = wnd.Dimensions.Y;
                         x -= AttachedScreen.Bounds.Left;
                         y -= AttachedScreen.Bounds.Top;
-                        var area = new Rectangle(x, y, w.Dimensions.Width - Form.BorderSize, w.Dimensions.Height - Form.BorderSize);
+                        var area = new Rectangle(x, y, wnd.Dimensions.Width - Form.BorderSize, wnd.Dimensions.Height - Form.BorderSize);
 
                         // Scale rectangles down to the thumbnail's desired size
                         var ar = Helpers.AspectRatioResize(new Size(AttachedScreen.Bounds.Width, AttachedScreen.Bounds.Height), 0, Form.PagerHeight);
@@ -92,7 +94,7 @@ namespace Switchie
 
                         area.X += Location.X;
                         area.Y += Location.Y;
-                        WindowAreas[w.Handle] = area;
+                        WindowAreas[wnd.Handle] = area;
 
                         // Window rectangle fill
                         g.FillRectangle(
@@ -102,9 +104,9 @@ namespace Switchie
                         // Window icon
                         var oldBounds = e.Graphics.ClipBounds;
                         e.Graphics.Clip = new Region(area);
-                        g.DrawImage(w.Icon, new Point(
-                            (area.X + area.Width / 2) - w.Icon.Width / 2,
-                            (area.Y + area.Height / 2) - w.Icon.Height / 2
+                        g.DrawImage(wnd.Icon, new Point(
+                            (area.X + area.Width / 2) - wnd.Icon.Width / 2,
+                            (area.Y + area.Height / 2) - wnd.Icon.Height / 2
                         ));
                         e.Graphics.Clip = new Region(oldBounds);